Query Korean census data by admin code (province or municipality) and year

The function queries a long format census data frame (censuskor) for specific administrative codes (if provided)

Usage

anycensusk(
  year = 2020,
  codes = NULL,
  type = c("population", "tax", "mortality", "economy"),
  level = c("adm1", "adm2")
)

Arguments

year

integer(1). One of 2010, 2015, or 2020.

codes

integer or character vector of admin codes (e.g. 11, 26) or admin names (e.g. "Seoul").

type

character(1). "population", "tax", "economy" or "mortality". Defaults to "population".

level

character(1). "adm1" for province-level or "adm2" for municipal-level. Defaults to "adm2".

Value

A data.frame object containing census data for the specified codes and year.

Examples

anycensusk(codes = "21", type = "mortality")
#> # A tibble: 16 × 8
#>     year adm1  adm2         adm2_code type      `All causes_total_p1p`
#>    <dbl> <chr> <chr>            <dbl> <chr>                      <dbl>
#>  1  2020 Busan Jung-gu          21010 mortality                   398.
#>  2  2020 Busan Seo-gu           21020 mortality                   395.
#>  3  2020 Busan Dong-gu          21030 mortality                   372.
#>  4  2020 Busan Yeongdo-gu       21040 mortality                   404.
#>  5  2020 Busan Busanjin-gu      21050 mortality                   332.
#>  6  2020 Busan Dongnae-gu       21060 mortality                   297.
#>  7  2020 Busan Nam-gu           21070 mortality                   311.
#>  8  2020 Busan Buk-gu           21080 mortality                   319.
#>  9  2020 Busan Haeundae-gu      21090 mortality                   302.
#> 10  2020 Busan Saha-gu          21100 mortality                   342.
#> 11  2020 Busan Geumjeong-gu     21110 mortality                   322.
#> 12  2020 Busan Gangseo-gu       21120 mortality                   290.
#> 13  2020 Busan Yeonje-gu        21130 mortality                   297.
#> 14  2020 Busan Suyeong-gu       21140 mortality                   294.
#> 15  2020 Busan Sasang-gu        21150 mortality                   363.
#> 16  2020 Busan Gijang-gun       21310 mortality                   329.
#> # ℹ 2 more variables: `All causes_male_p1p` <dbl>,
#> #   `All causes_female_p1p` <dbl>