Experimental function to load population data from KOSIS

Experimental function to load population data from KOSIS

Usage

load_population(year, cache = TRUE)

Arguments

year

The year for which to load the population data.

cache

Whether to cache the downloaded data. Default is TRUE.

Value

A data frame containing the population data for the specified year.

Note

This function is experimental and may change in future versions. It only supports loading raw population data in a single year. Reference table ID: DT_1IN1509

References

KOSIS

Examples

if (FALSE) { # \dontrun{
# Load population data for the year 2020
population_data <- load_population(2020)
} # }