South Korea Census Data

District level data including tax, population, business entities, and mortality in South Korea.

Usage

censuskor

Format

A data frame with 4410 rows and 9 variables:

adm1

Name of the province-level (Sido) administrative unit

adm2

Name of the district/municipal-level (Sigungu) administrative unit

adm2_code

Code of the district/municipal-level (Sigungu) administrative unit

type

Type of variable, e.g., "tax", "population", "mortality"

class1

First-level classification of the variable depending on the type

class2

Second-level classification of the variable depending on the type

unit

Unit of measurement for the variable

value

Value of the variable

year

Year of the census data, e.g., 2010, 2015, or 2020

Source

KOSIS (Korean Statistical Information Service)

Note

For temporal comparison, province names in adm1 field are standardized to the common names with no suffix in metropolitan cities and "-do" suffix in provinces. For example, "Seoul" instead of "Seoul Metropolitan City", and "Jeollabuk-do" instead of "Jeonbuk State".